+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [1.1] - 2021-12-22
+## [1.1] - 2021-12-28
 
 ### Added
 -Modal call to action that can display shortcodes and embed codes.
 -Modal dismiss button (fades out modals).
+-Add align-wide support.
 
 ### Changed
 -Fix header menu so items align to the right and don't stretch to fill the container along either axis.